add texture and scripts to the different gameObj
we are having problems with some of the scripts, not being add to the
wanted gameObjects when gitHub is sync. So when Ulrik updates, we have
to readd scripts and textures to the different gameObjects. This was
first done, Afterwards I created a new Enemy object called the
Deathstar, which was first modelled in Maya.